+++ stack_deployment_steps.go
@@ -29,7 +29,7 @@ type StackDeploymentStep struct {
 	UpdatedAt time.Time `jsonapi:"attr,created-at,iso8601"`
 
 	// Links
-	UploadUrl string `jsonapi:"links,upload-url,omitempty"`
+	UploadURL string `jsonapi:"links,upload-url,omitempty"`
 
 	// Relationships
 	StackDeploymentRun *StackDeploymentRun `jsonapi:"relation,stack-deployment-run"`