+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Unreleased
 
 ## Enhancements
-* Add the `UploadUrl` attribute to the `StackDeploymentStep` struct to support the deployment step GET endpoint by @shwetamurali [#1167](https://github.com/hashicorp/go-tfe/pull/1167)
+* Add the `UploadURL` attribute to the `StackDeploymentStep` struct to support the deployment step GET endpoint by @shwetamurali [#1167](https://github.com/h)
 
 # v1.88.0